Remove the changes to delegate list
